requirements.txt
librosa==0.10.2.post1
soundfile==0.12.1
streamlit==1.39.0
numpy==2.0.2
pandas==2.2.3
onnxruntime==1.19.2
aiocache==0.12.3
aiogram==3.16.0
docker==7.1.0
PyYAML==6.0.2

flake8==7.1.1
pyflakes==3.2.0
pycodestyle==2.12.1
pylint==3.3.1
mypy==1.13.0
radon==6.0.1
black==24.10.0
bandit==1.8.0